requirements.txt
mysql-connector-python>=8.0.0
openai>=1.0.0
python-dotenv>=0.19.0
typing>=3.7.4
asyncio>=3.4.3 